Remove unneeded workspace proc macro indirection crate
