+++ key-value-database/Cargo.toml
@@ -12,7 +12,7 @@ serde_json = "1"
 serde-wasm-bindgen = "0.6"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-sqlx = { version = "0.8", default-features = false, features = [ "runtime-tokio", "sqlite" ] }
+sqlx = { version = "0.8", default-features = false, features = [ "runtime-tokio", "sqlite", "derive" ] }
 
 [lints]
 workspace = true
